AI-ML-Internship-Eminds-feature-initial-setup/RAG_without_framework/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
python-dotenv==1.0.0
groq==0.4.2
